requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via
    #   kr-news (pyproject.toml)
    #   pydantic
anyio==4.9.0
    # via
    #   kr-news (pyproject.toml)
    #   httpx
    #   openai
certifi==2025.6.15
    # via
    #   kr-news (pyproject.toml)
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
diskcache==5.6.3
    # via llama-cpp-python
distro==1.9.0
    # via
    #   kr-news (pyproject.toml)
    #   openai
filelock==3.18.0
    # via huggingface-hub
fsspec==2025.5.1
    # via huggingface-hub
h11==0.16.0
    # via
    #   kr-news (pyproject.toml)
    #   httpcore
hf-xet==1.1.3
    # via huggingface-hub
httpcore==1.0.9
    # via
    #   kr-news (pyproject.toml)
    #   httpx
httpx==0.28.1
    # via
    #   kr-news (pyproject.toml)
    #   openai
huggingface-hub==0.33.0
    # via kr-news (pyproject.toml)
idna==3.10
    # via
    #   kr-news (pyproject.toml)
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via llama-cpp-python
jiter==0.10.0
    # via
    #   kr-news (pyproject.toml)
    #   openai
joblib==1.5.1
    # via
    #   kr-news (pyproject.toml)
    #   scikit-learn
llama-cpp-python==0.3.9
    # via kr-news (pyproject.toml)
markupsafe==3.0.2
    # via jinja2
numpy==2.3.0
    # via
    #   kr-news (pyproject.toml)
    #   llama-cpp-python
    #   pandas
    #   scikit-learn
    #   scipy
openai==1.86.0
    # via kr-news (pyproject.toml)
packaging==25.0
    # via huggingface-hub
pandas==2.3.0
    # via kr-news (pyproject.toml)
pydantic==2.11.7
    # via
    #   kr-news (pyproject.toml)
    #   openai
pydantic-core==2.33.2
    # via
    #   kr-news (pyproject.toml)
    #   pydantic
python-dateutil==2.9.0.post0
    # via
    #   kr-news (pyproject.toml)
    #   pandas
pytz==2025.2
    # via
    #   kr-news (pyproject.toml)
    #   pandas
pyyaml==6.0.2
    # via huggingface-hub
requests==2.32.4
    # via huggingface-hub
scikit-learn==1.7.0
    # via kr-news (pyproject.toml)
scipy==1.15.3
    # via
    #   kr-news (pyproject.toml)
    #   scikit-learn
setuptools==78.1.1
    # via kr-news (pyproject.toml)
six==1.17.0
    # via
    #   kr-news (pyproject.toml)
    #   python-dateutil
sniffio==1.3.1
    # via
    #   kr-news (pyproject.toml)
    #   anyio
    #   openai
threadpoolctl==3.6.0
    # via
    #   kr-news (pyproject.toml)
    #   scikit-learn
tqdm==4.67.1
    # via
    #   kr-news (pyproject.toml)
    #   huggingface-hub
    #   openai
typing-extensions==4.14.0
    # via
    #   kr-news (pyproject.toml)
    #   anyio
    #   huggingface-hub
    #   llama-cpp-python
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   kr-news (pyproject.toml)
    #   pydantic
tzdata==2025.2
    # via
    #   kr-news (pyproject.toml)
    #   pandas
urllib3==2.4.0
    # via requests
wheel==0.45.1
    # via kr-news (pyproject.toml)

rich~=14.0.0